+++ regimes/it/tax_categories.go
@@ -36,21 +36,11 @@ var categories = []*tax.CategoryDef{
 			i18n.EN: "Value Added Tax",
 			i18n.IT: "Imposta sul Valore Aggiunto",
 		},
+		Keys: tax.GlobalVATKeys(),
 		Rates: []*tax.RateDef{
 			{
-				Key: tax.RateZero,
-				Name: i18n.String{
-					i18n.EN: "Zero Rate",
-					i18n.IT: "Aliquota Zero",
-				},
-				Values: []*tax.RateValueDef{
-					{
-						Percent: num.MakePercentage(0, 3),
-					},
-				},
-			},
-			{
-				Key: tax.RateSuperReduced,
+				Keys: []cbc.Key{tax.KeyStandard},
+				Rate: tax.RateSuperReduced,
 				Name: i18n.String{
 					i18n.EN: "Minimum Rate",
 					i18n.IT: "Aliquota Minima",
@@ -62,7 +52,8 @@ var categories = []*tax.CategoryDef{
 				},
 			},
 			{
-				Key: tax.RateReduced,
+				Keys: []cbc.Key{tax.KeyStandard},
+				Rate: tax.RateReduced,
 				Name: i18n.String{
 					i18n.EN: "Reduced Rate",
 					i18n.IT: "Aliquota Ridotta",
@@ -74,7 +65,8 @@ var categories = []*tax.CategoryDef{
 				},
 			},
 			{
-				Key: tax.RateIntermediate,
+				Keys: []cbc.Key{tax.KeyStandard},
+				Rate: tax.RateIntermediate,
 				Name: i18n.String{
 					i18n.EN: "Intermediate Rate",
 					i18n.IT: "Aliquota Intermedia",
@@ -86,23 +78,16 @@ var categories = []*tax.CategoryDef{
 				},
 			},
 			{
-				Key: tax.RateStandard,
+				Keys: []cbc.Key{tax.KeyStandard},
+				Rate: tax.RateGeneral,
 				Name: i18n.String{
-					i18n.EN: "Ordinary Rate",
+					i18n.EN: "General Rate",
 					i18n.IT: "Aliquota Ordinaria",
 				},
 				Values: []*tax.RateValueDef{
 					{
 						Percent: num.MakePercentage(220, 3),
 					},
-				},
-			},
-			{
-				Key:    tax.RateExempt,
-				Exempt: true,
-				Name: i18n.String{
-					i18n.EN: "Exempt",
-					i18n.IT: "Esente",
 				},
 			},
 		},