+++ regimes/pt/invoices_test.go
@@ -38,7 +38,7 @@ func validInvoice() *bill.Invoice {
 				Taxes: tax.Set{
 					{
 						Category: "VAT",
-						Rate:     "standard",
+						Rate:     "general",
 					},
 				},
 			},